基于新的授权登陆机制， 初步搭建全部微服务体系， 初步搭建rocketmq与Spring Cloud Stream, 初步完成chat微服务 websocket的授权登录

+++ commons/src/main/java/top/camsyn/store/commons/entity/user/User.java
@@ -31,7 +31,7 @@ public class User extends Model<User> {
 
     private Integer credit;
 
-    private Double liyuan;
+    private Integer liyuan;
 
     private String headImage;
